html/parsing/html5lib/html-tree/tests20.dat

#data
<!doctype html><p><button><button>
#errors
button in button
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <button>
|       <button>

#data
<!doctype html><p><button><address>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <button>
|         <address>

#data
<!doctype html><p><button><blockquote>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <button>
|         <blockquote>

#data
<!doctype html><p><button><menu>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <button>
|         <menu>

#data
<!doctype html><p><button><p>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <button>
|         <p>

#data
<!doctype html><p><button><ul>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <button>
|         <ul>

#data
<!doctype html><p><button><h1>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <button>
|         <h1>

#data
<!doctype html><p><button><h6>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <button>
|         <h6>

#data
<!doctype html><p><button><listing>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <button>
|         <listing>

#data
<!doctype html><p><button><pre>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <button>
|         <pre>

#data
<!doctype html><p><button><form>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <button>
|         <form>

#data
<!doctype html><p><button><li>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <button>
|         <li>

#data
<!doctype html><p><button><dd>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <button>
|         <dd>

#data
<!doctype html><p><button><dt>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <button>
|         <dt>

#data
<!doctype html><p><button><plaintext>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <button>
|         <plaintext>

#data
<!doctype html><p><button><table>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <button>
|         <table>

#data
<!doctype html><p><button><hr>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <button>
|         <hr>

#data
<!doctype html><p><button><xmp>
#errors
EOF in xmp
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <button>
|         <xmp>

#data
<!doctype html><p><button></p>
#errors
no <p>
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       <button>
|         <p>

#data
<!doctype html><address><button></address>a
#errors
no </button>
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <address>
|       <button>
|     "a"

#data
<!doctype html><address><button></address>a
#errors
no </button>
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <address>
|       <button>
|     "a"

#data
<p><table></p>
#errors
no DOCTYPE
</p> in table
no </p>
EOF
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <p>
|       <table>

#data
<!doctype html><svg>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>

#data
<!doctype html><p><figcaption>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|     <figcaption>

#data
<!doctype html><p><summary>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|     <summary>

#data
<!doctype html><form><table><form>
#errors
form in table
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <form>
|       <table>

#data
<!doctype html><table><form><form>
#errors
form in table
form in table
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <form>

#data
<!doctype html><table><form></table><form>
#errors
form in table
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <table>
|       <form>

#data
<!doctype html><svg><foreignObject><p>
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg foreignObject>
|         <p>

#data
<!doctype html><svg><title>abc
#errors
EOF
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg title>
|         "abc"

#data
<option><span><option>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <option>
|       <span>
|         <option>

#data
<option><option>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <option>
|     <option>

#data
<math><annotation-xml><div>
#errors
no DOCTYPE
<div> in foreign
EOF
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|     <div>

#data
<math><annotation-xml encoding="application/svg+xml"><div>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|         encoding="application/svg+xml"
|     <div>

#data
<math><annotation-xml encoding="application/xhtml+xml"><div>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|         encoding="application/xhtml+xml"
|         <div>

#data
<math><annotation-xml encoding="aPPlication/xhtmL+xMl"><div>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|         encoding="aPPlication/xhtmL+xMl"
|         <div>

#data
<math><annotation-xml encoding="text/html"><div>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|         encoding="text/html"
|         <div>

#data
<math><annotation-xml encoding="Text/htmL"><div>
#errors
no DOCTYPE
EOF
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|         encoding="Text/htmL"
|         <div>

#data
<math><annotation-xml encoding=" text/html "><div>
#errors
no DOCTYPE
<div> in foreign
EOF
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math annotation-xml>
|         encoding=" text/html "
|     <div>
